DownloadApp Class
