+++ Cartfile.resolved
@@ -1,1 +1,1 @@
-github "PlayCover/PlayTools" "v2.0.2"
+github "PlayCover/PlayTools" "v2.0.3"
