+++ PlayCover/AppInstaller/Downloader.swift
@@ -6,93 +6,124 @@
 //
 
 import Foundation
+import DownloadManager
 
-func downloadApp(_ url: URL,
-                 _ app: StoreAppData,
-                 _ downloadVM: DownloadVM,
-                 _ warning: String?) {
+/// DownloaderManager can be configured through this struct, default values are as the same as below
+/// `public struct DownloadManagerConfig {`
+///    `public var maximumRetries = 3`
+///    `public var exponentialBackoffMultiplier = 10`
+///    `public var usesNotificationCenter = false`
+///    `public var showsLocalNotifications = false`
+///    `public var logVerbosity: LogVerbosity = .none
+/// `}
+///  Use `downloader.configuration = DownloadManagerConfig()` in
+///  `DownloadApp` class before `downloader.addDownload` to apply
+///  More detail: https://github.com/shapedbyiris/download-manager/blob/master/README.md
 
-    var observation: NSKeyValueObservation?
+class DownloadApp {
+    let url: URL
+    let app: StoreAppData
+    let warning: String?
 
-    if let warningMessage = warning {
-        let alert = NSAlert()
-        alert.messageText = NSLocalizedString(warningMessage, comment: "")
-        alert.informativeText = String(format: NSLocalizedString("ipaLibrary.alert.download",
-                                                                 comment: ""),
-                                       arguments: [app.name])
-        alert.alertStyle = .warning
-        alert.addButton(withTitle: NSLocalizedString("button.Yes", comment: ""))
-        alert.addButton(withTitle: NSLocalizedString("button.No", comment: ""))
-
-        if alert.runModal() == .alertSecondButtonReturn {
-            return
-        }
+    init(url: URL, app: StoreAppData, warning: String?) {
+        self.url = url
+        self.app = app
+        self.warning = warning
     }
 
-    if !downloadVM.downloading && !InstallVM.shared.installing {
-        lazy var urlSession = URLSession(configuration: .default, delegate: nil, delegateQueue: nil)
-        let downloadTask = urlSession.downloadTask(with: url, completionHandler: { url, urlResponse, error in
-            observation?.invalidate()
-            downloadComplete(url, app, downloadVM, urlResponse, error)
-        })
+    let dlVM = DownloadVM.shared
+    let inVM = InstallVM.shared
+    let downloader = DownloadManager.shared
 
-        observation = downloadTask.progress.observe(\.fractionCompleted) { progress, _ in
-            DispatchQueue.main.async {
-                downloadVM.progress = progress.fractionCompleted
+    func start() {
+        if let warningMessage = warning {
+            let alert = NSAlert()
+            alert.messageText = NSLocalizedString(warningMessage, comment: "")
+            alert.informativeText = String(
+                format: NSLocalizedString("ipaLibrary.alert.download", comment: ""),
+                arguments: [app.name]
+            )
+            alert.alertStyle = .warning
+            alert.addButton(withTitle: NSLocalizedString("button.Yes", comment: ""))
+            alert.addButton(withTitle: NSLocalizedString("button.No", comment: ""))
+
+            if alert.runModal() == .alertSecondButtonReturn {
+                return
+            } else {
+                proceedDownload()
             }
+        } else {
+            proceedDownload()
         }
-
-        downloadTask.resume()
-        downloadVM.downloading = true
-        downloadVM.progress = 0
-        downloadVM.storeAppData = app
-    } else {
-        Log.shared.error(PlayCoverError.waitDownload)
     }
-}
 
-func downloadComplete(_ url: URL?,
-                      _ app: StoreAppData,
-                      _ downloadVM: DownloadVM,
-                      _ urlResponse: URLResponse?,
-                      _ error: Error?) {
-    if error != nil {
-        Log.shared.error(error!)
+    func cancel() {
+        downloader.cancelAllDownloads()
+        self.dlVM.downloading = false
+        self.dlVM.progress = 0
+        self.dlVM.storeAppData = nil
     }
-    if let url = url {
-        var tmpDir: URL?
 
-        do {
-            tmpDir = try FileManager.default.url(for: .itemReplacementDirectory,
-                                                 in: .userDomainMask,
-                                                 appropriateFor: URL(fileURLWithPath: "/Users"),
-                                                 create: true)
+    private func proceedDownload() {
+        if dlVM.downloading && inVM.installing {
+            Log.shared.error(PlayCoverError.waitDownload)
+        } else {
+            let path = NSURL.fileURL(withPathComponents: [NSTemporaryDirectory(), app.bundleID])
+            downloader.addDownload(url: url,
+                                   destinationURL: path!,
+                                   onProgress: { progress in
+                self.dlVM.storeAppData = self.app
+                self.dlVM.downloading = true
+                // progress is a Float
+                self.dlVM.progress = Double(progress)
+            }, onCompletion: { error, fileURL in
+                guard error == nil else {
+                    self.dlVM.downloading = false
+                    self.dlVM.progress = 0
+                    self.dlVM.storeAppData = nil
+                    return Log.shared.error(error!)
+                }
+                self.dlVM.downloading = false
+                self.dlVM.progress = 0
+                self.proceedInstall(fileURL)
+            })
+        }
+    }
 
-            let tmpIpa = tmpDir!.appendingPathComponent(app.bundleID)
-                                .appendingPathExtension("ipa")
+    private func proceedInstall(_ url: URL?) {
+        self.dlVM.downloading = false
+        self.dlVM.progress = 0
+        if let url = url {
+            var tmpDir: URL?
+            do {
+                tmpDir = try FileManager.default.url(for: .itemReplacementDirectory,
+                                                     in: .userDomainMask,
+                                                     appropriateFor: URL(fileURLWithPath: "/Users"),
+                                                     create: true)
 
-            try FileManager.default.moveItem(at: url, to: tmpIpa)
-            uif.ipaUrl = tmpIpa
-            DispatchQueue.main.async {
+                let tmpIpa = tmpDir!.appendingPathComponent(app.bundleID)
+                                    .appendingPathExtension("ipa")
+
+                try FileManager.default.moveItem(at: url, to: tmpIpa)
+                uif.ipaUrl = tmpIpa
                 Installer.install(ipaUrl: uif.ipaUrl!, export: false, returnCompletion: { _ in
-                    FileManager.default.delete(at: tmpDir!)
-                    AppsVM.shared.apps = []
-                    AppsVM.shared.fetchApps()
-                    StoreVM.shared.resolveSources()
-                    NotifyService.shared.notify(
-                        NSLocalizedString("notification.appInstalled", comment: ""),
-                        NSLocalizedString("notification.appInstalled.message", comment: ""))
+                    DispatchQueue.main.async {
+                        FileManager.default.delete(at: tmpDir!)
+                        AppsVM.shared.apps = []
+                        AppsVM.shared.fetchApps()
+                        StoreVM.shared.resolveSources()
+                        NotifyService.shared.notify(
+                            NSLocalizedString("notification.appInstalled", comment: ""),
+                            NSLocalizedString("notification.appInstalled.message", comment: ""))
+                        self.dlVM.storeAppData = nil
+                    }
                 })
+            } catch {
+                if let tmpDir = tmpDir {
+                    FileManager.default.delete(at: tmpDir)
+                }
+                Log.shared.error(error)
             }
-        } catch {
-            if let tmpDir = tmpDir {
-                FileManager.default.delete(at: tmpDir)
-            }
-
-            Log.shared.error(error)
         }
     }
-    downloadVM.downloading = false
-    downloadVM.progress = 0
-//        downloadVM.storeAppData = nil
 }
